Fixed some compiler warnings
Lack of a default: clause on switch was causing a warning. Although "not possible" added a default clause and set the variable to zero that it was complaining about might have an undefined value. Also moved a variable declaration into the proper preprocessor section to eliminate a warning message that a variable was unused.